test cases for request and client level exclusions
