+++ sdk/cosmosdb/cosmos/src/ClientContext.ts
@@ -840,6 +840,8 @@ export class ClientContext {
         diagnosticNode,
         request.resourceType,
         request.operationType,
+        0,
+        options,
       );
       const response = await executePlugins(
         diagnosticNode,
